refactor: Optimization of the components Card and Cards
