+++ src/Pages/vehicles/Vehicles.jsx
@@ -5,7 +5,7 @@ import { useSelector } from "react-redux";
 import { ToastContainer } from "react-toastify";
 import { ButtonLoadMore } from "../../components";
 import { bodyCardVehicles } from "../../components/bodyCards/bodyCardVehicles";
-import { Cards } from "../../components/card/Cards";
+import { Cards } from "../../components/cards/Cards";
 
 import { GET_MORE_VEHICLES } from "../../redux/actions/actions";
 import { history } from "../../redux/reducers";
